crimson/os/seastore: move RecordSubmitter to segment_allocator.h
